+++ src/include/migraphx/target.hpp
@@ -124,6 +124,20 @@ supported_segments target_find_supported(T&, const_module_ref, support_metric)
     return {};
 }
 
+template <class T>
+value to_value_target(const T& x)
+{
+    return migraphx::to_value(x);
+}
+
+template <class T>
+void from_value_target(T& x, const value& v)
+{
+    if(not(v.is_object() or (v.empty() and v.is_array())))
+        MIGRAPHX_THROW("Value is not an object");
+    return migraphx::from_value(v, x);
+}
+
 #ifdef TYPE_ERASED_DECLARATION
 
 // Type-erased interface for:
@@ -143,6 +157,10 @@ struct MIGRAPHX_EXPORT target
     argument copy_from(const argument& input) const;
     // (optional)
     argument allocate(const shape& s) const;
+    // (optional)
+    value to_value() const;
+    // (optional)
+    void from_value(const value& v);
 };
 
 #else
@@ -213,6 +231,34 @@ struct target
         return target_allocate(private_detail_te_self, s);
     }
 
+    template <class T>
+    static auto private_detail_te_default_to_value(char, T&& private_detail_te_self)
+        -> decltype(private_detail_te_self.to_value())
+    {
+        return private_detail_te_self.to_value();
+    }
+
+    template <class T>
+    static value private_detail_te_default_to_value(float, T&& private_detail_te_self)
+    {
+        return to_value_target(private_detail_te_self);
+    }
+
+    template <class T>
+    static auto
+    private_detail_te_default_from_value(char, T&& private_detail_te_self, const value& v)
+        -> decltype(private_detail_te_self.from_value(v))
+    {
+        private_detail_te_self.from_value(v);
+    }
+
+    template <class T>
+    static void
+    private_detail_te_default_from_value(float, T&& private_detail_te_self, const value& v)
+    {
+        from_value_target(private_detail_te_self, v);
+    }
+
     template <class PrivateDetailTypeErasedT>
     struct private_te_unwrap_reference
     {
@@ -246,6 +292,11 @@ struct target
                  private_detail_te_default_allocate(char(0),
                                                     std::declval<PrivateDetailTypeErasedT>(),
                                                     std::declval<const shape&>()),
+                 private_detail_te_default_to_value(char(0),
+                                                    std::declval<PrivateDetailTypeErasedT>()),
+                 private_detail_te_default_from_value(char(0),
+                                                      std::declval<PrivateDetailTypeErasedT>(),
+                                                      std::declval<const value&>()),
                  void());
 
     template <class PrivateDetailTypeErasedT>
@@ -362,6 +413,18 @@ struct target
         return (*this).private_detail_te_get_handle().allocate(s);
     }
 
+    value to_value() const
+    {
+        assert((*this).private_detail_te_handle_mem_var);
+        return (*this).private_detail_te_get_handle().to_value();
+    }
+
+    void from_value(const value& v)
+    {
+        assert((*this).private_detail_te_handle_mem_var);
+        (*this).private_detail_te_get_handle().from_value(v);
+    }
+
     friend bool is_shared(const target& private_detail_x, const target& private_detail_y)
     {
         return private_detail_x.private_detail_te_handle_mem_var ==
@@ -383,6 +446,8 @@ struct target
         virtual argument copy_to(const argument& input) const                                   = 0;
         virtual argument copy_from(const argument& input) const                                 = 0;
         virtual argument allocate(const shape& s) const                                         = 0;
+        virtual value to_value() const                                                          = 0;
+        virtual void from_value(const value& v)                                                 = 0;
     };
 
     template <typename PrivateDetailTypeErasedT>
@@ -445,6 +510,18 @@ struct target
         {
 
             return private_detail_te_default_allocate(char(0), private_detail_te_value, s);
+        }
+
+        value to_value() const override
+        {
+
+            return private_detail_te_default_to_value(char(0), private_detail_te_value);
+        }
+
+        void from_value(const value& v) override
+        {
+
+            private_detail_te_default_from_value(char(0), private_detail_te_value, v);
         }
 
         PrivateDetailTypeErasedT private_detail_te_value;